requirements.txt
torch>=2.0.0
transformers>=4.30.0
PyQt5>=5.15.0
numpy>=1.24.0
onnxruntime-directml>=1.15.0
